backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
redis==5.2.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
tldextract==5.1.3
httpx==0.28.1
dnspython==2.7.0
python-whois==0.9.5